[Fix]: sidebar error resolve
(cherry picked from commit 92f2d3bdb94cb927708867c206f0daa91884b0c5)
